// apps/sales-copilot/src/generated/services/Crf5c_backgroundtasksService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Crf5c_backgroundtasksBase, Crf5c_backgroundtasks } from '../models/Crf5c_backgroundtasksModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Crf5c_backgroundtasksService {
  private static readonly dataSourceName = 'crf5c_backgroundtasks';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Crf5c_backgroundtasksBase, 'crf5c_backgroundtaskid'>): Promise<IOperationResult<Crf5c_backgroundtasks>> {
    const result = await Crf5c_backgroundtasksService.client.createRecordAsync<Omit<Crf5c_backgroundtasksBase, 'crf5c_backgroundtaskid'>, Crf5c_backgroundtasks>(
      Crf5c_backgroundtasksService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Crf5c_backgroundtasksBase, 'crf5c_backgroundtaskid'>>): Promise<IOperationResult<Crf5c_backgroundtasks>> {
    const result = await Crf5c_backgroundtasksService.client.updateRecordAsync<Partial<Omit<Crf5c_backgroundtasksBase, 'crf5c_backgroundtaskid'>>, Crf5c_backgroundtasks>(
      Crf5c_backgroundtasksService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Crf5c_backgroundtasksService.client.deleteRecordAsync(
      Crf5c_backgroundtasksService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Crf5c_backgroundtasks>> {
    const result = await Crf5c_backgroundtasksService.client.retrieveRecordAsync<Crf5c_backgroundtasks>(
      Crf5c_backgroundtasksService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Crf5c_backgroundtasks[]>> {
    const result = await Crf5c_backgroundtasksService.client.retrieveMultipleRecordsAsync<Crf5c_backgroundtasks>(
      Crf5c_backgroundtasksService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Crf5c_backgroundtasks> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Crf5c_backgroundtasksService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Crf5c_backgroundtasksService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}