// apps/sales-copilot/src/generated/models/Crf5c_backgroundtasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf5c_backgroundtasksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf5c_backgroundtasksstatecode = keyof typeof Crf5c_backgroundtasksstatecode;
export const Crf5c_backgroundtasksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf5c_backgroundtasksstatuscode = keyof typeof Crf5c_backgroundtasksstatuscode;

export interface Crf5c_backgroundtasksBase {
  crf5c_backgroundtaskid: string;
  crf5c_error?: string;
  crf5c_finishedon?: string;
  crf5c_name: string;
  crf5c_requestpayload?: string;
  crf5c_resultref?: string;
  crf5c_resultsummary?: string;
  crf5c_seenon?: string;
  crf5c_startedon?: string;
  crf5c_status?: string;
  crf5c_targetentityid?: string;
  crf5c_targetentitytype?: string;
  crf5c_targetname?: string;
  crf5c_tasktype?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf5c_backgroundtasksstatecode;
  statuscode?: Crf5c_backgroundtasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf5c_backgroundtasks extends Crf5c_backgroundtasksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}